spread: add support for system specific "flags" and use in qemu
This commit adds a concept of system specific flags. This is useful
to enable certain properties on the given system. The first use-case
is to switch virtio on for a specific system. This will be useful
for UC20 where the UEFI grub bootloader is very very slow (minutes)
when loading the kernel in non-virtio mode.

The flags concept will probably be useful for other use-cases as
well. One example is to add "uefi" to google backend so that
systems that have this flag are created with UEFI enabled.